myapp/api/serializers.py
from rest_framework import serializers
from base.models import Item
from todo.models import Todo

class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = ('__all__')


class TodoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Todo
        fields = ('__all__')

